requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

chromadb==0.5.5
gravis==0.1.0
ipython==8.25.0
langchain-huggingface==0.0.3
nest_asyncio==1.6.0
networkx==3.3
pandas==2.2.2
pydantic==2.9.2
pydot==3.0.1
python-dotenv==1.0.1
pyvis==0.3.2
scipy==1.13.1
sentence-transformers==2.7.0
tqdm==4.66.4
